Move Scene conversion tests
